[MS] Refactoring of HomePage
